+++ src/components/logo/logo.tsx
@@ -1,0 +1,13 @@
+import { Link } from 'react-router-dom';
+
+export default function Logo(): React.JSX.Element {
+  return (
+    <div className="logo">
+      <Link to="/" className="logo__link">
+        <span className="logo__letter logo__letter--1">W</span>
+        <span className="logo__letter logo__letter--2">T</span>
+        <span className="logo__letter logo__letter--3">W</span>
+      </Link>
+    </div>
+  );
+}
