+++ src/components/private-route/PrivateRoute.tsx
@@ -1,0 +1,16 @@
+import { Navigate } from 'react-router-dom';
+import { AppRoute } from '../../enums/AppRoute.ts';
+
+type PrivateRouteProps = {
+  children: JSX.Element;
+};
+
+function PrivateRoute(props: PrivateRouteProps): JSX.Element {
+  const { children } = props;
+
+  const isAuth = false;
+
+  return isAuth ? children : <Navigate to={AppRoute.Login} />;
+}
+
+export default PrivateRoute;
