+++ src/enums/AppRoute.ts
@@ -1,0 +1,8 @@
+export const enum AppRoute {
+  Main = '/',
+  Login = '/login',
+  MyList = '/mylist',
+  Films = '/films',
+  Review = ':id/review',
+  Player = '/player/:id'
+}
