+++ src/enums/AuthorizationStatus.ts
@@ -1,0 +1,5 @@
+export const enum AuthorizationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unknown = 'UNKNOWN',
+}
