+++ src/pages/head-guest/head-guest.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
+import Logo from '../../components/logo/logo';
 
 function HeadGuest(): React.FunctionComponent {
   return (
@@ -9,17 +11,11 @@ function HeadGuest(): React.FunctionComponent {
       </div>
       <h1 className="visually-hidden">WTW</h1>
       <header className="page-header">
-        <div className="logo">
-          <a className="logo__link">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </a>
-        </div>
+        <Logo />
         <div className="user-block">
-          <a href="sign-in.html" className="user-block__link">
+          <Link to="/login" className="user-block__link">
             Sign in
-          </a>
+          </Link>
         </div>
       </header>
     </section>
