+++ src/pages/movie-page-details/movie-page-details.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
+import Logo from '../../components/logo/logo';
+import { Link } from 'react-router-dom';
 
-function MoviePage(): React.FunctionComponent {
+function MoviePageDetails(): React.FunctionComponent {
   return (
     <>
       <section className="film-card film-card--full">
@@ -13,13 +15,7 @@ function MoviePage(): React.FunctionComponent {
           </div>
           <h1 className="visually-hidden">WTW</h1>
           <header className="page-header film-card__head">
-            <div className="logo">
-              <a href="main.html" className="logo__link">
-                <span className="logo__letter logo__letter--1">W</span>
-                <span className="logo__letter logo__letter--2">T</span>
-                <span className="logo__letter logo__letter--3">W</span>
-              </a>
-            </div>
+            <Logo />
             <ul className="user-block">
               <li className="user-block__item">
                 <div className="user-block__avatar">
@@ -49,7 +45,7 @@ function MoviePage(): React.FunctionComponent {
                   type="button"
                 >
                   <svg viewBox="0 0 19 19" width={19} height={19}>
-                    <use xlinkHref="#play-s" />
+                    <use xlinkHref="#play-s"/>
                   </svg>
                   <span>Play</span>
                 </button>
@@ -58,7 +54,7 @@ function MoviePage(): React.FunctionComponent {
                   type="button"
                 >
                   <svg viewBox="0 0 19 20" width={19} height={20}>
-                    <use xlinkHref="#add" />
+                    <use xlinkHref="#add"/>
                   </svg>
                   <span>My list</span>
                   <span className="film-card__count">9</span>
@@ -83,53 +79,71 @@ function MoviePage(): React.FunctionComponent {
             <div className="film-card__desc">
               <nav className="film-nav film-card__nav">
                 <ul className="film-nav__list">
-                  <li className="film-nav__item film-nav__item--active">
-                    <a href="src/components/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
+                  <li className="film-nav__item">
+                    <a href="src/pages/movie-page-details/movie-page-details#" className="film-nav__link">
                       Overview
                     </a>
                   </li>
-                  <li className="film-nav__item">
-                    <a href="src/components/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
+                  <li className="film-nav__item film-nav__item--active">
+                    <a href="src/pages/movie-page-details/movie-page-details#" className="film-nav__link">
                       Details
                     </a>
                   </li>
                   <li className="film-nav__item">
-                    <a href="src/components/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
+                    <a href="src/pages/movie-page-details/movie-page-details#" className="film-nav__link">
                       Reviews
                     </a>
                   </li>
                 </ul>
               </nav>
-              <div className="film-rating">
-                <div className="film-rating__score">8,9</div>
-                <p className="film-rating__meta">
-                  <span className="film-rating__level">Very good</span>
-                  <span className="film-rating__count">240 ratings</span>
-                </p>
-              </div>
-              <div className="film-card__text">
-                <p>
-                  In the 1930s, the Grand Budapest Hotel is a popular European
-                  ski resort, presided over by concierge Gustave H. (Ralph
-                  Fiennes). Zero, a junior lobby boy, becomes Gustave`s friend
-                  and protege.
-                </p>
-                <p>
-                  Gustave prides himself on providing first-class service to the
-                  hotel`s guests, including satisfying the sexual needs of the
-                  many elderly women who stay there. When one of Gustave`s
-                  lovers dies mysteriously, Gustave finds himself the recipient
-                  of a priceless painting and the chief suspect in her murder.
-                </p>
-                <p className="film-card__director">
-                  <strong>Director: Wes Anderson</strong>
-                </p>
-                <p className="film-card__starring">
-                  <strong>
-                    Starring: Bill Murray, Edward Norton, Jude Law, Willem Dafoe
-                    and other
-                  </strong>
-                </p>
+              <div className="film-card__text film-card__row">
+                <div className="film-card__text-col">
+                  <p className="film-card__details-item">
+                    <strong className="film-card__details-name">
+                      Director
+                    </strong>
+                    <span className="film-card__details-value">
+                      Wes Anderson
+                    </span>
+                  </p>
+                  <p className="film-card__details-item">
+                    <strong className="film-card__details-name">
+                      Starring
+                    </strong>
+                    <span className="film-card__details-value">
+                      Bill Murray, <br/>
+                      Edward Norton, <br/>
+                      Jude Law, <br/>
+                      Willem Dafoe, <br/>
+                      Saoirse Ronan, <br/>
+                      Tony Revoloru, <br/>
+                      Tilda Swinton, <br/>
+                      Tom Wilkinson, <br/>
+                      Owen Wilkinson, <br/>
+                      Adrien Brody, <br/>
+                      Ralph Fiennes, <br/>
+                      Jeff Goldblum
+                    </span>
+                  </p>
+                </div>
+                <div className="film-card__text-col">
+                  <p className="film-card__details-item">
+                    <strong className="film-card__details-name">
+                      Run Time
+                    </strong>
+                    <span className="film-card__details-value">1h 39m</span>
+                  </p>
+                  <p className="film-card__details-item">
+                    <strong className="film-card__details-name">Genre</strong>
+                    <span className="film-card__details-value">Comedy</span>
+                  </p>
+                  <p className="film-card__details-item">
+                    <strong className="film-card__details-name">
+                      Released
+                    </strong>
+                    <span className="film-card__details-value">2014</span>
+                  </p>
+                </div>
               </div>
             </div>
           </div>
@@ -203,11 +217,11 @@ function MoviePage(): React.FunctionComponent {
         </section>
         <footer className="page-footer">
           <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
+            <Link to="/" className="logo__link logo__link--light">
               <span className="logo__letter logo__letter--1">W</span>
               <span className="logo__letter logo__letter--2">T</span>
               <span className="logo__letter logo__letter--3">W</span>
-            </a>
+            </Link>
           </div>
           <div className="copyright">
             <p>© 2019 What to watch Ltd.</p>
@@ -218,4 +232,4 @@ function MoviePage(): React.FunctionComponent {
   );
 }
 
-export default MoviePage;
+export default MoviePageDetails;
