+++ src/pages/movie-page-reviews/movie-page-reviews.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import Logo from '../../components/logo/logo';
+import {Link} from 'react-router-dom';
 
 function MoviePageReviews(): React.FunctionComponent {
   return (
@@ -13,13 +15,7 @@ function MoviePageReviews(): React.FunctionComponent {
           </div>
           <h1 className="visually-hidden">WTW</h1>
           <header className="page-header film-card__head">
-            <div className="logo">
-              <a href="main.html" className="logo__link">
-                <span className="logo__letter logo__letter--1">W</span>
-                <span className="logo__letter logo__letter--2">T</span>
-                <span className="logo__letter logo__letter--3">W</span>
-              </a>
-            </div>
+            <Logo/>
             <ul className="user-block">
               <li className="user-block__item">
                 <div className="user-block__avatar">
@@ -49,7 +45,7 @@ function MoviePageReviews(): React.FunctionComponent {
                   type="button"
                 >
                   <svg viewBox="0 0 19 19" width={19} height={19}>
-                    <use xlinkHref="#play-s" />
+                    <use xlinkHref="#play-s"/>
                   </svg>
                   <span>Play</span>
                 </button>
@@ -58,7 +54,7 @@ function MoviePageReviews(): React.FunctionComponent {
                   type="button"
                 >
                   <svg viewBox="0 0 19 20" width={19} height={20}>
-                    <use xlinkHref="#add" />
+                    <use xlinkHref="#add"/>
                   </svg>
                   <span>My list</span>
                   <span className="film-card__count">9</span>
@@ -84,17 +80,17 @@ function MoviePageReviews(): React.FunctionComponent {
               <nav className="film-nav film-card__nav">
                 <ul className="film-nav__list">
                   <li className="film-nav__item">
-                    <a href="#" className="film-nav__link">
+                    <a href="src/pages/movie-page-reviews/movie-page-reviews#" className="film-nav__link">
                       Overview
                     </a>
                   </li>
                   <li className="film-nav__item">
-                    <a href="#" className="film-nav__link">
+                    <a href="src/pages/movie-page-reviews/movie-page-reviews#" className="film-nav__link">
                       Details
                     </a>
                   </li>
                   <li className="film-nav__item film-nav__item--active">
-                    <a href="#" className="film-nav__link">
+                    <a href="src/pages/movie-page-reviews/movie-page-reviews#" className="film-nav__link">
                       Reviews
                     </a>
                   </li>
@@ -281,11 +277,11 @@ function MoviePageReviews(): React.FunctionComponent {
         </section>
         <footer className="page-footer">
           <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
+            <Link to="/" className="logo__link logo__link--light">
               <span className="logo__letter logo__letter--1">W</span>
               <span className="logo__letter logo__letter--2">T</span>
               <span className="logo__letter logo__letter--3">W</span>
-            </a>
+            </Link>
           </div>
           <div className="copyright">
             <p>© 2019 What to watch Ltd.</p>
