+++ src/pages/movie-page/movie-page.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
+import Logo from '../../components/logo/logo';
+import {Link} from 'react-router-dom';
 
-function MoviePageInList(): React.FunctionComponent {
+function MoviePage(): React.FunctionComponent {
   return (
     <>
       <section className="film-card film-card--full">
@@ -13,13 +15,7 @@ function MoviePageInList(): React.FunctionComponent {
           </div>
           <h1 className="visually-hidden">WTW</h1>
           <header className="page-header film-card__head">
-            <div className="logo">
-              <a href="main.html" className="logo__link">
-                <span className="logo__letter logo__letter--1">W</span>
-                <span className="logo__letter logo__letter--2">T</span>
-                <span className="logo__letter logo__letter--3">W</span>
-              </a>
-            </div>
+            <Logo/>
             <ul className="user-block">
               <li className="user-block__item">
                 <div className="user-block__avatar">
@@ -49,7 +45,7 @@ function MoviePageInList(): React.FunctionComponent {
                   type="button"
                 >
                   <svg viewBox="0 0 19 19" width={19} height={19}>
-                    <use xlinkHref="#play-s" />
+                    <use xlinkHref="#play-s"/>
                   </svg>
                   <span>Play</span>
                 </button>
@@ -57,15 +53,15 @@ function MoviePageInList(): React.FunctionComponent {
                   className="btn btn--list film-card__button"
                   type="button"
                 >
-                  <svg viewBox="0 0 18 14" width={18} height={14}>
-                    <use xlinkHref="#in-list" />
+                  <svg viewBox="0 0 19 20" width={19} height={20}>
+                    <use xlinkHref="#add"/>
                   </svg>
                   <span>My list</span>
                   <span className="film-card__count">9</span>
                 </button>
-                <a href="add-review.html" className="btn film-card__button">
+                <Link to="films:id/review" className="btn film-card__button">
                   Add review
-                </a>
+                </Link>
               </div>
             </div>
           </div>
@@ -84,17 +80,17 @@ function MoviePageInList(): React.FunctionComponent {
               <nav className="film-nav film-card__nav">
                 <ul className="film-nav__list">
                   <li className="film-nav__item film-nav__item--active">
-                    <a href="#" className="film-nav__link">
+                    <a href="src/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
                       Overview
                     </a>
                   </li>
                   <li className="film-nav__item">
-                    <a href="#" className="film-nav__link">
+                    <a href="src/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
                       Details
                     </a>
                   </li>
                   <li className="film-nav__item">
-                    <a href="#" className="film-nav__link">
+                    <a href="src/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
                       Reviews
                     </a>
                   </li>
@@ -149,9 +145,9 @@ function MoviePageInList(): React.FunctionComponent {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link className="small-film-card__link" to="films:id">
                   Fantastic Beasts: The Crimes of Grindelwald
-                </a>
+                </Link>
               </h3>
             </article>
             <article className="small-film-card catalog__films-card">
@@ -164,9 +160,9 @@ function MoviePageInList(): React.FunctionComponent {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link className="small-film-card__link" to="/films:id">
                   Bohemian Rhapsody
-                </a>
+                </Link>
               </h3>
             </article>
             <article className="small-film-card catalog__films-card">
@@ -179,9 +175,9 @@ function MoviePageInList(): React.FunctionComponent {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link className="small-film-card__link" to="/films:id">
                   Macbeth
-                </a>
+                </Link>
               </h3>
             </article>
             <article className="small-film-card catalog__films-card">
@@ -194,20 +190,20 @@ function MoviePageInList(): React.FunctionComponent {
                 />
               </div>
               <h3 className="small-film-card__title">
-                <a className="small-film-card__link" href="film-page.html">
+                <Link className="small-film-card__link" to="/films:id">
                   Aviator
-                </a>
+                </Link>
               </h3>
             </article>
           </div>
         </section>
         <footer className="page-footer">
           <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
+            <Link to="/" className="logo__link logo__link--light">
               <span className="logo__letter logo__letter--1">W</span>
               <span className="logo__letter logo__letter--2">T</span>
               <span className="logo__letter logo__letter--3">W</span>
-            </a>
+            </Link>
           </div>
           <div className="copyright">
             <p>© 2019 What to watch Ltd.</p>
@@ -218,4 +214,4 @@ function MoviePageInList(): React.FunctionComponent {
   );
 }
 
-export default MoviePageInList;
+export default MoviePage;
